requirements.txt
requests>=2.25.0
pyyaml>=5.4.0

